pyproject.toml
[build-system]
name = "poetry-core"
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "openjailbreak"
version = "0.1.0"
description = "A comprehensive framework for automatic LLM jailbreak research"
authors = ["zenithfang"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/zenithfang/OpenJailbreak"
repository = "https://github.com/zenithfang/OpenJailbreak"
documentation = "https://github.com/zenithfang/OpenJailbreak/blob/main/docs/"

[tool.poetry.dependencies]
python = "^3.8"
datasets = "^2.14.0"
huggingface-hub = "^0.19.0"
litellm = "^1.72.0"
torch = "^2.0.0"
transformers = "^4.36.0"
vllm = {version = "^0.2.0", optional = true}
pandas = "^2.0.0"
matplotlib = "^3.7.0"
tqdm = "^4.66.0"
numpy = "^1.24.0"
nltk = "^3.8.1"
regex = "^2023.6.3"
peft = "^0.5.0"
colorama = "^0.4.6"
nanogcg = "^0.1.0"
openai = "^1.3.0"
click = "^8.1.3"
pyyaml = "^6.0"

[tool.poetry.extras]
vllm = ["vllm"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.1.0"
black = "^23.3.0"
isort = "^5.12.0"
ruff = "^0.0.278"
pre-commit = "^3.3.3"

[tool.black]
line-length = 100
target-version = ["py38"] 